profile missing image

+++ app/Http/Controllers/API/UserController.php
@@ -7,15 +7,19 @@ use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Support\Facades\Hash;
 
+
 class UserController extends Controller
 {
+
     /**
      * Create a new controller instance.
      *
      * @return void
      */
     public function __construct()
-    { }
+    {
+        $this->middleware('auth:api');
+    }
 
     /**
      * Display a listing of the resource.
@@ -24,11 +28,10 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
-        // $this->authorize('isAdmin');
-        // if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
-        //     return User::latest()->paginate(5);
-        // }
+        $this->authorize('isAdmin');
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            return User::latest()->paginate(5);
+        }
     }
 
     /**
@@ -87,7 +90,6 @@ class UserController extends Controller
             $request->merge(['password' => Hash::make($request['password'])]);
         }
 
-
         $user->update($request->all());
         return ['message' => "Success"];
     }
@@ -118,15 +120,12 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-
         $user = User::findOrFail($id);
-
         $this->validate($request, [
             'name' => 'required|string|max:191',
             'email' => 'required|string|email|max:191|unique:users,email,' . $user->id,
             'password' => 'sometimes|min:6'
         ]);
-
         $user->update($request->all());
         return ['message' => 'Updated the user info'];
     }
@@ -139,20 +138,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-
-        // $this->authorize('isAdmin');
-
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
-        // delete the user
-
         $user->delete();
-
         return ['message' => 'User Deleted'];
     }
 
     public function search()
     {
-
         if ($search = \Request::get('q')) {
             $users = User::where(function ($query) use ($search) {
                 $query->where('name', 'LIKE', "%$search%")
@@ -161,7 +154,6 @@ class UserController extends Controller
         } else {
             $users = User::latest()->paginate(5);
         }
-
         return $users;
     }
 }
